Refactor: Introduce folderId to GridItem and update dependencies
This commit introduces a `folderId: String?` field to the `GridItem` data class. This change necessitates updates across various domain and data layer components:

- **Domain Models & Use Cases:**
    - `GridItem` now includes `folderId`.
    - Use cases like `AddPinShortcutToHomeScreenUseCase`, `AddPinWidgetToHomeScreenUseCase`, and `UpdateGridItemsUseCase` have been updated to initialize or handle the new `folderId` field.
    - `FolderGridItemAlgorithm` has been modified, and its return type adjusted.

- **Data Layer:**
    - Mappers (`ApplicationInfoGridItemMapper`, `WidgetGridItemMapper`, `ShortcutInfoGridItemMapper`, `DefaultFolderGridItemRepository`) are updated to correctly map the `folderId`.
    - `EblanDatabase` schema (`1.json`) has been updated to reflect the new `folderId` column in `ApplicationInfoGridItemEntity`, `WidgetGridItemEntity`, and `ShortcutInfoGridItemEntity`.
    - `FolderGridItemWrapperEntity` has been removed.

- **Feature Layer:**
    - Various screens and helpers within the `feature/home` and `feature/pin` modules (e.g., `PagerScreenHelper`, `WidgetScreen`, `ShortcutScreen`, `ApplicationScreen`, `PinViewModel`, `PinScreen`) have been updated to accommodate the `folderId` in `GridItem` and `GridItemSource`.
    - `ExperimentalCoroutinesApi` annotation has been added to `HomeScreen`, `HomeViewModel`, and `GetHomeDataUseCase`.
    - `DragScreen` and `ResizeScreen` have TODOs added for handling `GridItemData.Folder`.

- **Build & Configuration:**
    - Domain modules (`domain/framework`, `domain/grid`, `domain/repository`, `domain/use-case`, `domain/model`, `domain/common`) now use the `com.eblan.launcher.jvmLibrary` plugin instead of `com.eblan.launcher.library`.
    - Android library desugaring and related configurations (`androidDesugarJdkLibs`) have been removed.
    - `kotlin-stdlib` dependency updated.
    - Kotlin Compose plugin removed from `AndroidLibraryConventionPlugin` and `AndroidApplicationConventionPlugin`.
    - Removed snapshot repository from `settings.gradle.kts`.
    - Removed `consumer-proguard-rules.pro` from `domain/model`.
    - Removed offline mode from `.idea/gradle.xml`.
    - `KotlinAndroid.kt` updated to remove desugaring and experimental coroutines API flag.
