+++ .idea/gradle.xml
@@ -58,6 +58,5 @@
         </option>
       </GradleProjectSettings>
     </option>
-    <option name="offlineMode" value="true" />
   </component>
 </project>
