+++ build-logic/convention/src/main/kotlin/AndroidApplicationConventionPlugin.kt
@@ -32,8 +32,8 @@ class AndroidApplicationConventionPlugin : Plugin<Project> {
         with(target) {
             with(pluginManager) {
                 apply("com.android.application")
+                apply("org.jetbrains.kotlin.android")
                 apply("com.eblan.launcher.lint")
-                apply("org.jetbrains.kotlin.plugin.compose")
             }
 
             extensions.configure<ApplicationExtension> {
