+++ data/repository/src/main/kotlin/com/eblan/launcher/data/repository/mapper/ApplicationInfoGridItemMapper.kt
@@ -9,6 +9,7 @@ import com.eblan.launcher.domain.model.GridItemData
 internal fun ApplicationInfoGridItemEntity.asGridItem(): GridItem {
     return GridItem(
         id = id,
+        folderId = folderId,
         page = page,
         startRow = startRow,
         startColumn = startColumn,
