+++ domain/grid/build.gradle.kts
@@ -17,12 +17,7 @@
  */
 
 plugins {
-    alias(libs.plugins.com.eblan.launcher.library)
-    alias(libs.plugins.com.eblan.launcher.libraryJacoco)
-}
-
-android {
-    namespace = "com.eblan.launcher.domain.grid"
+    alias(libs.plugins.com.eblan.launcher.jvmLibrary)
 }
 
 dependencies {
