--- domain/model/consumer-proguard-rules.pro
@@ -1,1 +1,0 @@
--keep class com.android.geto.domain.model.** { *; }
