+++ domain/model/src/main/kotlin/com/eblan/launcher/domain/model/GridItem.kt
@@ -2,6 +2,7 @@ package com.eblan.launcher.domain.model
 
 data class GridItem(
     val id: String,
+    val folderId: String?,
     val page: Int,
     val startRow: Int,
     val startColumn: Int,
