+++ domain/use-case/src/main/kotlin/com/eblan/launcher/domain/usecase/AddPinWidgetToHomeScreenUseCase.kt
@@ -14,16 +14,14 @@ import kotlinx.coroutines.flow.first
 import kotlinx.coroutines.withContext
 import java.io.File
 import javax.inject.Inject
-import kotlin.uuid.ExperimentalUuidApi
-import kotlin.uuid.Uuid
 
 class AddPinWidgetToHomeScreenUseCase @Inject constructor(
     private val gridCacheRepository: GridCacheRepository,
     private val userDataRepository: UserDataRepository,
     private val fileManager: FileManager,
 ) {
-    @OptIn(ExperimentalUuidApi::class)
     suspend operator fun invoke(
+        id: String,
         className: String,
         componentName: String,
         configure: String?,
@@ -59,8 +57,6 @@ class AddPinWidgetToHomeScreenUseCase @Inject constructor(
                 fileManager.widgetsDirectory,
                 className,
             ).absolutePath
-
-            val id = Uuid.random().toHexString()
 
             val gridHeight = rootHeight - dockHeight
 
@@ -107,6 +103,7 @@ class AddPinWidgetToHomeScreenUseCase @Inject constructor(
 
             val gridItem = GridItem(
                 id = id,
+                folderId = null,
                 page = initialPage,
                 startRow = 0,
                 startColumn = 0,
