+++ feature/home/src/main/kotlin/com/eblan/launcher/feature/home/screen/drag/DragScreen.kt
@@ -353,6 +353,8 @@ private fun GridItemContent(
                         )
                     }
                 }
+
+                is GridItemData.Folder -> TODO()
             }
         }
     }
