+++ feature/home/src/main/kotlin/com/eblan/launcher/feature/home/screen/widget/WidgetScreen.kt
@@ -287,6 +287,7 @@ fun getWidgetGridItem(
 
     return GridItem(
         id = id,
+        folderId = null,
         page = page,
         startRow = 0,
         startColumn = 0,
