+++ settings.gradle.kts
@@ -10,9 +10,6 @@ pluginManagement {
         }
         mavenCentral()
         gradlePluginPortal()
-        maven {
-            url = uri("https://androidx.dev/snapshots/builds/13787427/artifacts/repository")
-        }
     }
 }
 dependencyResolutionManagement {
@@ -20,9 +17,6 @@ dependencyResolutionManagement {
     repositories {
         google()
         mavenCentral()
-        maven {
-            url = uri("https://androidx.dev/snapshots/builds/13787427/artifacts/repository")
-        }
     }
 }
